+++ packages/api/test/car.spec.js
@@ -1,7 +1,7 @@
 /* eslint-env mocha */
 import assert from 'assert'
 import { CID } from 'multiformats/cid'
-import { sha256 } from 'multiformats/hashes/sha2'
+import { sha256, sha512 } from 'multiformats/hashes/sha2'
 import * as pb from '@ipld/dag-pb'
 import { CarWriter } from '@ipld/car'
 import fetch, { Blob } from '@web-std/fetch'
@@ -9,6 +9,7 @@ import { endpoint, clusterApi, clusterApiAuthHeader } from './scripts/constants.
 import { createCar } from './scripts/car.js'
 import { MAX_BLOCK_SIZE } from '../src/constants.js'
 import { getTestJWT } from './scripts/helpers.js'
+import { PIN_OK_STATUS } from '../src/utils/pin.js'
 
 describe('POST /car', () => {
   it('should add posted CARs to Cluster', async () => {
@@ -41,8 +42,8 @@ describe('POST /car', () => {
 
     const statusRes = await fetch(new URL(`status/${cid}`, endpoint))
     const status = await statusRes.json()
-    const pinned = status.pins.find(pin => pin.status === 'Pinned')
-    assert(pinned, 'CID is Pinned')
+    const pinInfo = status.pins.find(pin => PIN_OK_STATUS.includes(pin.status))
+    assert(pinInfo, `status is one of ${PIN_OK_STATUS}`)
 
     const clusterPeersRes = await fetch(new URL('peers', clusterApi), {
       headers: {
@@ -51,7 +52,7 @@ describe('POST /car', () => {
     })
     const clusterPeers = await clusterPeersRes.json()
     // assert that peerId from the status belongs to one of the cluster ipfs nodes.
-    assert(clusterPeers.some(peer => peer.ipfs.id === pinned.peerId))
+    assert(clusterPeers.some(peer => peer.ipfs.id === pinInfo.peerId))
   })
 
   it('should throw for blocks bigger than the maximum permitted size', async () => {
@@ -207,5 +208,69 @@ describe('POST /car', () => {
     assert.strictEqual(res.ok, false)
     const { message } = await res.json()
     assert.strictEqual(message, 'Invalid CAR file received: CAR must contain at least one non-root block')
+  })
+
+  it('should allow a CAR with unsupported hash function', async () => {
+    const token = await getTestJWT('test-upload', 'test-upload')
+
+    const bytes = pb.encode({ Data: new Uint8Array(), Links: [] })
+    // we dont support sha512 yet!
+    const hash = await sha512.digest(bytes)
+    const cid = CID.create(1, pb.code, hash)
+
+    const { writer, out } = CarWriter.create([cid])
+    writer.put({ cid, bytes })
+    writer.close()
+
+    const carBytes = []
+    for await (const chunk of out) {
+      carBytes.push(chunk)
+    }
+
+    const res = await fetch(new URL('car', endpoint), {
+      method: 'POST',
+      headers: {
+        Authorization: `Bearer ${token}`,
+        'Content-Type': 'application/car'
+      },
+      body: new Blob(carBytes)
+    })
+
+    assert(res, 'Server responded')
+    assert(res.ok, 'Server response ok')
+    const resBody = await res.json()
+    assert(resBody.cid, 'Server response payload has `cid` property')
+    assert.strictEqual(resBody.cid, cid.toString(), 'Server responded with expected CID')
+  })
+
+  it('should throw for CAR with a block where the bytes do match the CID', async () => {
+    const token = await getTestJWT('test-upload', 'test-upload')
+
+    const bytes = pb.encode({ Data: new Uint8Array(), Links: [] })
+    const hash = await sha256.digest(bytes)
+    const cid = CID.create(1, pb.code, hash)
+
+    const { writer, out } = CarWriter.create([cid])
+    bytes[bytes.length - 1] = bytes[bytes.length - 1] + 1 // mangle a byte
+    writer.put({ cid, bytes })
+    writer.close()
+
+    const carBytes = []
+    for await (const chunk of out) {
+      carBytes.push(chunk)
+    }
+
+    const res = await fetch(new URL('car', endpoint), {
+      method: 'POST',
+      headers: {
+        Authorization: `Bearer ${token}`,
+        'Content-Type': 'application/car'
+      },
+      body: new Blob(carBytes)
+    })
+
+    assert.strictEqual(res.ok, false)
+    const { message } = await res.json()
+    assert.strictEqual(message, `Invalid CAR file received: block data does not match CID for ${cid.toString()}`)
   })
 })