+++ packages/db/test/pin-request.spec.js
@@ -23,6 +23,7 @@ describe('pin-request', () => {
       location: {
         peerId: '12D3KooWFe387JFDpgNEVCP5ARut7gRkX7YuJCXMStpkq714ziK6',
         peerName: 'web3-storage-sv15',
+        ipfsPeerId: '12D3KooWR19qPPiZH4khepNjS3CLXiB7AbrbAD4ZcDjN1UjGUNE2',
         region: 'region'
       }
     },
@@ -31,6 +32,7 @@ describe('pin-request', () => {
       location: {
         peerId: '12D3KooWFe387JFDpgNEVCP5ARut7gRkX7YuJCXMStpkq714ziK7',
         peerName: 'web3-storage-sv16',
+        ipfsPeerId: '12D3KooWR19qPPiZH4khepNjS3CLXiB7AbrbAD4ZcDjN1UjGUNE3',
         region: 'region'
       }
     }